requirements.txt
fastapi==0.54.1
motor==2.1.0
pytest==5.4.1
pytest-asyncio==0.11.0
factory-boy==2.12.0
